Move forked PatternTokenizer back to original namespace, comments
