+++ src/main/java/org/elasticsearch/index/analysis/PatternTokenizerFactory.java
@@ -20,6 +20,7 @@
 package org.elasticsearch.index.analysis;
 
 import org.apache.lucene.analysis.Tokenizer;
+import org.apache.lucene.analysis.pattern.XPatternTokenizer;
 import org.elasticsearch.ElasticsearchIllegalArgumentException;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.inject.assistedinject.Assisted;
@@ -51,6 +52,6 @@ public class PatternTokenizerFactory extends AbstractTokenizerFactory {
 
     @Override
     public Tokenizer create(Reader reader) {
-        return new PatternTokenizer(reader, pattern, group);
+        return new XPatternTokenizer(reader, pattern, group);
     }
 }
